+++ cmd/projects/list.go
@@ -129,12 +129,15 @@ func List() error {
 		close(data)
 	}()
 
-	results := sort.FromChannel(data, &sort.Options{
+	results, err := sort.FromChannel(data, &sort.Options{
 		OrderBy:    orderBy,
 		SortBy:     sortBy,
 		GroupBy:    groupBy,
 		StructType: gitlab.Project{},
 	})
+	if err != nil {
+		return err
+	}
 
 	w := tabwriter.NewWriter(os.Stdout, 0, 0, 1, ' ', tabwriter.TabIndent)
 	fmt.Fprintf(w, "COUNT\tREPOSITORY\tHOSTS\tCACHED\n")
