+++ cmd/users/block.go
@@ -97,10 +97,13 @@ func Block() error {
 		close(blocked)
 	}()
 
-	results := sort.FromChannel(blocked, &sort.Options{
+	results, err := sort.FromChannel(blocked, &sort.Options{
 		OrderBy:    []string{blockBy},
 		StructType: gitlab.User{},
 	})
+	if err != nil {
+		return err
+	}
 
 	w := tabwriter.NewWriter(os.Stdout, 0, 0, 1, ' ', tabwriter.TabIndent)
 	fmt.Fprintf(w, "COUNT\tUSER\tHOSTS\tCACHED\n")
